rg_display: Removed unecessary optimization in ili9341_set_window that caused glitches
Sometimes our knowledge of the window position and the LCD's actual position seem to get out of sync. It might be caused by a number of things but for now we'll just send the full coordinates on every set_window calls...

This adds an average of 4-12 4 bytes spi transactions per frame, which isn't too bad.