Replaced miniz with zlib to hopefully improve stability and compatibility
Compression in zlib uses 3x more memory than our modified miniz.
Decompression is similar and flash usage is only a couple KB more.

So it will not be able to compress a screenshot when running larger roms. However, unlike miniz, it handles out of memory gracefully and its stack usage is much less.

We can also work around that in the future by reclaiming emulation caches or framebuffers, etc.

It's also the reference implementation so it will hopefully fix all our issues...