PCE: Significant performance improvement
By removing the tile/sprite cache we can fit more things into the ESP32's internal RAM.

I've optimized the planning of sprite/tiles as much as I can. In the end, with the cache removed, there was a 5% slow down.

But with all that freed memory it was possible to move one of the framebuffers (and some smaller things) into internal RAM resulting in performance improvements of up to 40%!