Removed the SPI mutex for ODROID-GO
The SPI locking is now done at a lower level, in the sdspi driver.

My first attempt (to avoid having to patch esp-idf) was to scan the memory to find the spi_device_handle and do the locking in sdcard_do_transaction. It worked, but it feels wrong.

Since we already patch that exact to improve SD Card compat, I think it's cleaner to just add the two locking lines to it...