DOOM: Revamped menu code
It was very hard to follow and work it. It went from 5100 lines to 2300. (Though many comments need to be added back...)

Many bugs will have to be worked out I'm sure. That was likely pass 1 out of 2 or 3.

The combined setup menus allowed me to reduce the size of the built-in PRBOOM.WAD as well (30KB saved!)