retro-go: Added an internal implementation of printf
The newlib version (both std and nano) does allocations for most modifiers and it also fairly slow.

This is problematic when I want to get messages out despite a corrupted or full heap.

The code is from mpaland/printf[1] but I have modified the public api/wrappers a little.
#include <rg_printf.h> needs to be added to use it in apps.

It isn't included in rg_system because I haven't tested all printf usage by emulators with it.

1. https://github.com/mpaland/printf/