PCE: Fixed out-of-bound accesses in save state code
Timer variables are defined as SVAR_4 but they are 16bits, causing out-of-bound accesses during save/load.

It might be more appropriate to have reload/counter/running as uint8_t, but this break save state format so it will wait until we refactor save state code...

The size doesn't affect behavior anyway, we constrain the values and we rely on underflow arithmetic which is the same for 8/16/32.