rg_audio: Added locking when accessing the hardware
In some apps (pce-go, doom, snes) the audio is running in a separate task.

Which means that we might call rg_audio_submit in the middle of a call to rg_audio_set_sink, resulting in a crash or weird noises.

This change adds a mutex to be used by init/deinit/submit.