rg_storage: No longer support the older /odroid/data save location
A message will be shown to the user on first boot, telling them what to do.

This change is because go-play and retro-go have diverged somewhat in save state format and it regularly causes corruption when going back and forth.

This changes only affects ODROID-GO, other devices already used /retro-go/saves.