rg_audio: rg_audio_submit is now constant, it won't modify the source buffer
We used to resample directly over the old buffer, now we copy to a new buffer(on the stack, 720 Bytes) as we resample it.

One downside is that we now call i2s_write more often because our local buffer is small (typically 3 calls per frame instead of 2).

This should not be an issue as long as we keep the size of our local buffer in sync with the I2S DMA buffer...