rg_network: Added network support
rg_network is responsible for establishing and maintaining a connection to an access point, or to a peer when running in AP mode.

It provides a simple interface: One local IP to represent us, One remote IP to represent our peer (there can be only one).

This commit also draws a cute radio icon when it's connecting and connected