Moved retro-core's components to the retro-core folder instead of global
It makes it a bit simpler to use lto when the code is contained in the same directory.

It's also simpler to build a single object for experimenting.

Not sure it's worth the disruption hence why it's a separate branch for now, but it might be merged to dev/master...