rg_system: Use rg_system_tick to track responsiveness
Before we used the last read_gamepad time to judge if the app was still alive. It was done that way because input is always polled even during blocking retro-go things (dialogs, wait for key) whereas the tick counter is not running during those operations. So it was simpler to implement.

But now both rg_gui_dialog and rg_input_wait_for_key are tick-aware and we can rely on ticking to feed the watchdog.