Launcher: Added option to start a temporary access point (amend)
This starts a temporary access point (must be started manually every time). A permanent access point can be configured in wifi.json if desired.

The use of a password is annoying, but someone might transfer their wifi.json over the temp AP and it could be captured by someone if left open... Unlikely, but still.