SNES: Second attempt at fixing controls dialog crash ( #71)
I think the problem is that the options array isn't zeroed anymore so the spacer (second option set) contains bogus pointers to label and value and this crash rg_gui_dialog...

Solution is to set label and value and update_cb. I've also made the array local to the loop for good measure.