GBC: Added a setting to not use system time for the RTC
Normally, after loading a save state, we reload the RTC with the system time.

This can cause problems when the user manually adjust the RTC in-game or in the menu and expects it to be restored precisely when reloading a state.

I've added a hidden (for now) setting SysTime that when set to 0 will never overwrite the RTC.