rg_display: Fixed bug causing some buffers to go unused
In spi_queue_transaction() we automatically copy the data to an spi buffer if it's too large but we didn't set the flag for it to be returned after the transaction.

Two commands in the init sequence caused this to happen, meaning we were effectively always running with SPI_BUFFER_COUNT-2 buffers.

The bug is now fixed, but i have also changed SPI_BUFFER_COUNT from 6 to 5, as we were evidently running fine with 4...