rg_display: Reverted some of f1900c5d to fix glitches
`lcd_send_buffer` was causing some display glitches. Presumably because of the fact that it tracks its own window and we got out of sync with the LCD.

I've fixed the issue in `lcd_send_buffer` (by removing the internal cursor) but it is now less performant than an explicit `lcd_set_window`+`lcd_send_data`, so I reverted that for now.