retro-go: removed rg_task_delete
rg_task_delete isn't that useful as it doesn't allow clean termination of tasks (and on many platforms it doesn't even terminate the task/thread).

I replaced it with a wrapper that will cleanup when the task returns.

Not ideal to have a wrapper but it shouldn't have any impact on performance.