NES: Added a memory block to back unmapped regions
The CPU and PPU do direct memory access but it's not impossible for the region they try to access to be unmapped.

Since it's direct, we can't trap it.

I added dummy memory to trap it.

I also made ppu_setpage/getpage consistent with mem_setpage/getpage so it's less confusing.