NES: Added sound emulation to mapper Namco 163 (iNES 19)
I mostly followed the nesdev wiki and cross checked some values with FCEUX. My implementation is not very accurate (and neither is FCEUX's, so I should have used a better reference...) but it's nice to finally have music in Pacman Championship Edition!

Save state format had to be slightly modified but it doesn't break compatibility.