rg_storage: scandir now uses a callback system to save memory
Previously on large lists (2000+) it could fail because no contiguous block large enough could be allocated.

I tried a few methods to work around it (make multiple smaller allocations, use a bucket system, use a linked list).

But they all had the issue that they had to be freed correctly + they were slower + they caused fragmentation.

Now it no longer returns an array but takes a callback to run on every file. Which makes it a tad more annoying to use, but oh well.

On the bright side, filenames are no longer limited to 90 characters!