rg_gui: Added hidden entry support + Allow draw buffer to grow
We keep an internal buffer to draw to before blitting to the display. This buffer used to be width * 20. Trying to draw any more pixels would trash memory.

Our text drawing routine in particular draws an entire line at a time, which can be larger with bigger fonts.

I should make it able to draw half a line at a time if needed, but it seemed better to allow the buffer to grow for all drawing routines.