rg_display+rg_system: Removed update mode setting and improved frame pacing
Now that partial update is no longer glitchy, there is no reason to ever disable it.

However, the `Update` setting was also indirectly doing one more thing: It imposed a frameskip of 1 when `Full` was selected.

This did prevent stuttering in some games.

But that's an indirect effect, and even I keep forgetting that it's the reason why it can counter-intuitively improve performance.

So I think it's better to handle the frameskip directly instead.