Fixed in-game frameskip doing dumb stuff on unexpected delays
All our games have the same primitive frameskip system that I wrote forever ago. It isn't very clever, but it did the job.

Now, consider the following:
`skipFrames = (elapsed + frameTime / 2) / frameTime;`

If the game stalled for 5 seconds because it's loading something from SD, it would result in skipFrames being around 160.

This is almost 3 seconds of no updating for a one time event.

This isn't good. And with the MSX emulator, it's outright unusable if emulating a floppy.

The in-game frame skip can now only be 1.

Anything more will have to come from the auto-frameskip.