Added fMSX 6.0 emulator
The emulator source is unmodified but it is dynamically patched at compile time (by forcing an include file).

Both keyboard and joystick are emulated. For keyboard I looked up the most used keys in games. With feedback I'll be able to tweak that, but the virtual keyboard also works if other keys are needed.

Save states work through the in-game menu but not yet through retro-go's menu.

Launcher images are also missing.