rg_display: Improved scaling
This new scaling use floating point and can resize in any direction and to any size.

Still have more testing to do to make sure it's not a regression...

Because it looks worse in some games. But I think that's just a coincidence caused by the fact that we double different lines, the scaling isn't wrong. It looks better in others and the same in most.

I think improving the filtering or adding an integer scaling mode would help.