Fix margin to have a black bar instead of garbage on the metal case
This unit came in two case variations. The plastic case, which was modified from a simple LCD keyring toy game and for which the unit was designed to fit inside, and a metal case that was designed specifically to house the internals.

Turns out, the viewing window on the metal case is higher than on the plastic case, presumably to make better use of the actual screen. With the previous margin, garbage is shown on the bottom of the screen if playing with a metal case. With the plastic case, the bottom of the screen is physically masked off to prevent light bleed, which is why I didn't notice before. Extending the bottom margin prevents the garbage on the bottom of the metal units.