DOOM: Prefer external RAM for all allocations (#109)
DOOM2 tends to use the entirety of the internal RAM, leaving nothing for DMA operations such as SD card access.

This patch makes DOOM prefer to use external RAM. It will only start using internal when it has filled the external RAM.

Ideally we should locate critical structs in DOOM and put them manually in internal RAM, for performance.

But that is a lot of work and the game remains playable with this patch, I do not think there are any downsides.