Launcher: Reduced memory usage
Currently all the images we load are cached after they're first loaded. As we've added more and more tabs over the years, the memory usage has become unreasonable. After a user has cycled through all the tabs, we have very small amounts of memory left, causing all kinds of stability issues.

This commit now frees background images as the tab is left, as they're the biggest offenders by far.

But this causes slower tab switching, so in the future we should either:

- Always cache a compressed copy of the unloaded images with a very fast compression algo
- Or at least cache the raw PNG data when it is loaded from SD Card...