Launcher: Allow manual crc cache building
This allow the user to build up the CRC32 cache to make cover art work better.

Previously an idle task was used to accomplish this, but it gave no feedback as to progress and didn't work when connected to wifi.

I think this is a better solution, even though the new menu entry might confuse some users as to its purpose. (I should probably add some description on the screen...)