Add RG_STORAGE_FALLBACK_FLASH mechanism
If RG_STORAGE_FALLBACK_FLASH is defined, and retro-go
hasn't mounted any other storage, it will fallback to
trying to find an internal flash partion labelled 'vfs'
and mount it. It will also format it if unformatted.

As this code is only compiled in if RG_STORAGE_FALLBACK_FLASH
is defined or RG_STORAGE_DRIVER == 4, it won't increase the
filesize unnecessarily.