rg_storage: Add RG_STORAGE_FALLBACK_FLASH mechanism (#129)
* Add RG_STORAGE_FALLBACK_FLASH mechanism

If RG_STORAGE_FALLBACK_FLASH is defined, and retro-go
hasn't mounted any other storage, it will fallback to
trying to find an internal flash partion labelled 'vfs'
and mount it. It will also format it if unformatted.

As this code is only compiled in if RG_STORAGE_FALLBACK_FLASH
is defined or RG_STORAGE_DRIVER == 4, it won't increase the
filesize unnecessarily.

* Rename RG_STORAGE_FALLBACK_FLASH to RG_STORAGE_INTERNAL_FLASH_FALLBACK

Technically, an SD card is also flash memory, just external.
So it makes sense to add an explicit "INTERNAL" modifier to the FLASH.

Also, changed to word order to make more sense.

* Improve comment

Technically, an SD card is also flash memory, just external.
So it makes sense to add an explicit "Internal" modifier to the word "Flash".