rg_input: inclusive ADC ranges
On the Fri3D Camp badge, the ADC readings for moving
the analog joystick to the left range from 0 to 1024.

But when those values are placed in the RG_GAMEPAD_ADC_MAP,
moving the joystick to the left doesn't do anything, because
the code is checking 0 > 0, which is false.

Moving the joystick slightly left does work,
because then the ADC readout will be slightly above 0.

The fix is to make the configured ADC values inclusive,
so the code will check that 0 >= 0, which is true.