DOOM: Always play audio at 22050
It was running at 11025 for performance reasons (and that's the native sound effects samplerate) but it can cause issues with some I2S DACs and in esp-idf 5.x it causes issue with the internal DAC as well.

To compensate for the additional overhead I optimized the audio loop to avoid unnecessary work.