Launcher: No longer show cover errors
To distinguish between a missing file and a broken file we issue a `rg_storage_exists()` call.

This causes an extra SD card transaction which takes 5-10ms.

I recently moved the exists check to happen only on load failure, avoiding the transaction in some cases.

But I don't think making the distinction in the GUI is worth the extra 5-10ms of blockage.

Maybe in the future I'll make `rg_surface_load_image_file` return an error code, somehow, that could be used.