Updater: allow longer update names to be shown
The update file names for the Fri3D Camp 2024 Badge are a bit longer than
32 bytes, as we like to have a long, proper description and also an
"OVERWRITES_EVERYTHING" warning to the .img files that contain a fresh
copy of the pre-populated FAT storage partition, so that the user knows
the storage (including configuration and games) will be overwritten.

Therefore, 32 bytes was a bit short for the update names,
and this commit doubles it to 64 bytes.

Of course, if for some reason, a project wants to keep the names short,
they can still give them short names, and then the retro-go updater GUI
will only show those short names. It's only *if* you want a longer, more
descriptive name to be shown, you can now, with this change.