rg_tool: Define all tool paths at the start of the script (#146)
All the tools we execute are now defined at the start of the script and can be overridden if needed.

It also defines the absolute path to gen_esp32part.py which seems to be missing in some environments.

(cherry picked from commit 7ceefba72a3eb0beef66f44056b7d000299ae8bf)