i2s audio driver: fix conflict with GPIO0 due to default 0 for .mck_io_num
In the i2s_pin_config_t that's passed to i2s_set_pin, .mck_io_num was not
initialized to GPIO_NUM_NC, so it defaulted to 0. Which matches the GPIO that
the Fri3D Camp 2024 Badge is using for the START button, resulting in "START"
being pressed constantly, which caused it to loop through the emulators.