Add a bit of HTTP client debug information
This is useful to know what's going on without having to dig deep.

For instance, the retro-go HTTP API will return 404 to:
http://192.168.4.1/%2fsd%2fcrash.log
but not to:
http://192.168.4.1/sd%2fcrash.log

But without seeing the status code, it's kind of guess work
as to why exactly rg_network_http_open() failed.

Or it will not provide a Content-Length field for download file requests,
but it *will* assume that other HTTP servers do provide it, such as in
updater.c's download_file() to show the total bytes to download.

An informational "content length: 0" would have helped to understand this.