rg_gui: Added function `rg_gui_draw_message`
This function draws a message in the center of the screen.

In many places I've used `rg_gui_draw_dialog` to show a non-blocking message on top of the screen.

This new function does the same thing in a more explicit manner. It also supports multi line messages and takes in printf format, which will be handy since it's often used to show status or progress.