rg_system: Added message queue to rg_task
Having a built-in messaging system for all tasks is very convenient.

Tasks are free to ignore it or to use it to receive work units. There is no longer a need for separate queues!

This also allows tasks to receive and react to some standard messages like STOP/SHUTDOWN.

An additional benefit is that this will be very easy to make it work in SDL2 compared to full fledged queues. (I have already made a very unsafe and innefficient implementation to showcase, but it needs more SDL_Mutex/SDL_Cond...).