rg_tool: Create dummy partition table to avoid build failures on larger apps
The default partition table has a 1M partition for the app and there is an option for a 1.5M. If the app binary ends up being larger, the build will fail.

Now rg_tool.py will create a dummy partition table with a 3M partition which should allow all builds to succeed.