rg_system: Avoid unnecessary flash write in rg_system_switch_app
Now that the launcher is part of retro-core, switching between it and most emulators means rebooting to the same partition.

esp_ota_set_boot_partition doesn't check if changes happened, it always overwrites the OTA data partition.

This patch checks if the current partition is ok, and if so do not call esp_ota_set_boot_partition.