rg_display: Prevent flicker on forced redraw
Previously the screen was cleared when `rg_display_force_redraw` is called and scaling isn't set to fullscreen.

That is because the viewport won't redraw everything so garbage could remain on the sides.

Now instead of checking for scaling mode, we check if the viewport covers everything.

If it does, we don't bother blanking the screen.

This prevents an annoying flicker when leaving the menu in the launcher.