Updated fri3d-2024 sdkconfig to fix CPU frequency (#170)
Ensure the CPU is set to 240Mhz.

The value names as of esp-idf 5.x should be: `CONFIG_ESP_DEFAULT_CPU_FREQ_MHZ_X`

But `CONFIG_ESP32S3_DEFAULT_CPU_FREQ_X` seems to work for both 4.4 and 5.x, so I went with that for now.

The previous `CONFIG_ESP32_DEFAULT_CPU_FREQ_X` was just not working on 5.x, always defaulting to 160Mhz.