Launcher: Updater now has to be explicitly enabled by the target
It made no sense to have it enabled for all targets when it couldn't work without support from the factory app.

Now if a target doesn't define a download location and updater app then it will only do a version check, not download the update.

The updater/version check can now also be entirely disabled by a target.