NES: Fixed bogus corrections in rom_loadmem when crc32 isn't available (#181)
When loading a ROM, retro-go calcs the crc32 of the rom and searches for it in the database. If the CRC32 is 0, the last entry in the database will be selected, which is invalid.

This patch just confirms that the selected entry is valid.

A proper patch might be to not search the database at all if we don't have a crc function. But that would change multiple lines versus just this one!