Cleaning up ODROID-GO's config.h to hopefully make it *the* reference/template
I want it to contain all supported options (commented out) so that it can be used as a source of truth and template.

It would be nice to rename some of those constants to be nicer and more consistent by section, but it's not worth breaking out-of-tree targets for now...

Once I'm satisfied with the formatting I'll update the other targets to match.