Removed esplay-s3 target, I cannot find information about it
I thought it was: https://github.com/FantasyGmm/esplay_micro_s3_hardware

But the schematics does not match the pinouts in config.h...

No more info in the PR itself.

So I guess this was a prototype, which isn't terribly useful to us if nobody has one or can build one.