Enable the psram-cache-issue fix for all components
Previously it was only enabled in DOOM and the launcher because they write to PSRAM a lot, triggering the bug often.

Other emulators tend to use it only to store their ROM in PSRAM, so the write bug doesn't usually occur. Which is why the fix was disabled until now.

But it definitely causes instability regardless. In this commit it is now enabled for all emulators, but not for the esp-idf side or libretro-go.

This should be a reasonable compromise. More testing is needed to see if it makes any emulator unusably slow...

So far I'm seeing 10-15% higher CPU usage which is... not great. Thankfully most emulators were at 50-60% usage, so there's headroom...