rg_display+ili9341: RG_SCREEN_ROTATION now actually does something
It's simply impossible to intuitively map human-readable units (degrees or mirroring) to the actual reflected effect on all devices. There's too much hardware variability that we have to support.

So what I did is make a small range of valid values and the user should try them one by one until it looks fine.

Not ideal, but there's only 7 values to try...