rg_system: Adjustable message queue depth when creating a new task
Tasks in retro-go wait for messages and work on them as they come. Before today the queue size/depth was fixed to 1, meaning that if the task was busy working on a message, any subsequent call to rg_task_send would block.

It sucks that so many files need to be modified but defining the queue size in rg_task_create is the way to go, not adding a rg_task_resize_queue() or something like that (IMHO).